requirements.txt
astroid==3.3.10
black==25.1.0
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
dill==0.4.0
et_xmlfile==2.0.0
ff3==1.0.2
flake8==7.3.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
tomlkit==0.13.3
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xlrd==2.0.2
